+++ waspc/examples/todoApp/package.json
@@ -7,16 +7,18 @@
   },
   "dependencies": {
     "@playwright/test": "^1.50.0",
+    "chai": "^5.2.0",
     "linebyline": "^1.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.26.2",
-    "wasp": "file:.wasp/out/sdk/wasp",
-    "wait-port": "^1.1.0"
+    "wait-port": "^1.1.0",
+    "wasp": "file:.wasp/out/sdk/wasp"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.10",
     "@tailwindcss/typography": "^0.5.16",
+    "@types/chai": "^5.2.2",
     "@types/express": "^5.0.0",
     "@types/react": "^18.0.37",
     "@types/uuid": "^9.0.8",
